Revert "[WFLY-9989] Bump messaging-activemq model version to 3.1.0"
This reverts commit 41c9b4583953649376908e8710d6f7f8011735c9.
