+++ messaging-activemq/src/main/resources/subsystem-templates/messaging-activemq.xml
@@ -2,7 +2,7 @@
 <!--  See src/resources/configuration/ReadMe.txt for how the configuration assembly works -->
 <config default-supplement="default">
     <extension-module>org.wildfly.extension.messaging-activemq</extension-module>
-    <subsystem xmlns="urn:jboss:domain:messaging-activemq:3.1">
+    <subsystem xmlns="urn:jboss:domain:messaging-activemq:3.0">
         <server name="default">
             <?CLUSTERED?>
