--- messaging-activemq/src/test/resources/org/wildfly/extension/messaging/activemq/subsystem_3_1_ha-policy.xml
@@ -1,93 +1,0 @@
-
-<subsystem xmlns="urn:jboss:domain:messaging-activemq:3.1">
-    <server name="ha-policy-live-only-scale-down-discovery-group">
-        <live-only>
-            <scale-down enabled="${scale-down.enabled:true}"
-                        cluster-name="${scale-down.cluster-name:mycluster}"
-                        group-name="${scale-down.group-name:mygroup}"
-                        discovery-group="groupC"/>
-        </live-only>
-    </server>
-    <server name="ha-policy-live-only-scale-down-connectors">
-        <live-only>
-            <scale-down enabled="${scale-down.enabled:true}"
-                        cluster-name="${scale-down.cluster-name:mycluster}"
-                        group-name="${scale-down.group-name:mygroup}"
-                        connectors="netty in-vm"/>
-        </live-only>
-    </server>
-    <server name="ha-policy-replication-master">
-        <replication-master
-                cluster-name="${replication-master.cluster-name:mycluster}"
-                group-name="${replication-master.cluster-name:mygroup}"
-                check-for-live-server="${replication-master.check-for-live-server:false}"
-                initial-replication-sync-timeout="${replication-master.initial-replication-sync-timeout:1234}"/>
-    </server>
-    <server name="ha-policy-replication-slave">
-        <replication-slave
-                cluster-name="${replication-slave.cluster-name:mycluster}"
-                group-name="${replication-slave.cluster-name:mygroup}"
-                allow-failback="${replication-slave.allow-failback:true}"
-                initial-replication-sync-timeout="${replication-master.initial-replication-sync-timeout:1234}"
-                restart-backup="${replication-slave.restart-backup:true}"
-                max-saved-replicated-journal-size="${replication-slave.max-saved-replicated-journal-size:24}">
-            <scale-down enabled="${replication-slave-scale-down.enabled:true}"
-                        cluster-name="${replication-slave-scale-down.cluster-name:mycluster}"
-                        group-name="${replication-slave-scale-down.group-name:mygroup}"
-                        connectors="netty"/>
-        </replication-slave>
-    </server>
-    <server name="ha-policy-replication-colocated">
-        <replication-colocated request-backup="${replication-colocated.request-backup:false}"
-                               backup-request-retries="${replication-colocated.backup-request-retries:-1}"
-                               backup-request-retry-interval="${replication-colocated.backup-request-retry-interval:5098}"
-                               max-backups="${replication-colocated.max-backups:5}"
-                               backup-port-offset="${replication-colocated.backup-port-offset:500}"
-                               excluded-connectors="netty">
-            <master cluster-name="${replication-colocated-master.cluster-name:mycluster}"
-                    group-name="${replication-colocated-master.cluster-name:mygroup}"
-                    check-for-live-server="${replication-colocated-master.check-for-live-server:false}" />
-            <slave cluster-name="${replication-colocated-slave.cluster-name:mycluster}"
-                   group-name="${replication-colocated-slave.cluster-name:mygroup}"
-                   allow-failback="${replication-colocated-slave.allow-failback:true}"
-                   initial-replication-sync-timeout="${replication-colocated-slave.initial-replication-sync-timeout:1234}"
-                   restart-backup="${replication-colocated-slave.restart-backup:true}"
-                   max-saved-replicated-journal-size="${replication-colocated-slave.max-saved-replicated-journal-size:24}">
-                <scale-down enabled="${replication-colocated-slave-scale-down.enabled:true}"
-                            cluster-name="${replication-colocated-slave-scale-down.cluster-name:mycluster}"
-                            group-name="${replication-colocated-slave-scale-down.group-name:mygroup}"
-                            connectors="netty"/>
-            </slave>
-        </replication-colocated>
-    </server>
-    <server name="ha-policy-shared-store-master">
-        <shared-store-master failover-on-server-shutdown="${shared-store-master.failover-on-server-shutdown:true}" />
-    </server>
-    <server name="ha-policy-shared-store-slave">
-        <shared-store-slave allow-failback="${shared-store-slave.allow-failback:false}"
-                            failover-on-server-shutdown="${shared-store-slave.failover-on-server-shutdown:true}"
-                            restart-backup="${shared-store-slave.restart-backup:false}">
-            <scale-down enabled="${shared-store-slave-scale-down.enabled:true}"
-                        cluster-name="${shared-store-slave-scale-down.cluster-name:mycluster}"
-                        group-name="${shared-store-slave-slave-scale-down.group-name:mygroup}"
-                        connectors="netty" />
-        </shared-store-slave>
-    </server>
-    <server name="ha-policy-shared-store-colocated">
-        <shared-store-colocated request-backup="${shared-store-colocated.request-backup:false}"
-                                backup-request-retries="${shared-store-colocated.backup-request-retries:-1}"
-                                backup-request-retry-interval="${shared-store-colocated.backup-request-retry-interval:5098}"
-                                max-backups="${shared-store-colocated.max-backups:5}"
-                                backup-port-offset="${shared-store-colocated.backup-port-offset:500}">
-            <master failover-on-server-shutdown="${shared-store-colocated-master.failover-on-server-shutdown:true}" />
-            <slave allow-failback="${shared-store-colocated-slave.allow-failback:false}"
-                   failover-on-server-shutdown="${shared-store-colocated-slave.failover-on-server-shutdown:true}"
-                   restart-backup="${shared-store-colocated-slave.restart-backup:false}">
-                <scale-down enabled="${shared-store-colocated-slave-scale-down.enabled:true}"
-                            cluster-name="${shared-store-colocated-slave-scale-down.cluster-name:mycluster}"
-                            group-name="${shared-store-colocated-slave-slave-scale-down.group-name:mygroup}"
-                            connectors="netty" />
-            </slave>
-        </shared-store-colocated>
-    </server>
-</subsystem>
